Split ACP runtime manager tests
