+++ scripts/file-length-baseline.json
@@ -13,7 +13,6 @@
   "src/backend/services/ratchet/service/ratchet.service.ts": 1035,
   "src/backend/services/run-script/service/run-script.service.test.ts": 2332,
   "src/backend/services/run-script/service/run-script.service.ts": 1026,
-  "src/backend/services/session/service/acp/acp-runtime-manager.test.ts": 2705,
   "src/backend/services/session/service/acp/acp-runtime-manager.ts": 1449,
   "src/backend/services/session/service/acp/codex-app-server-adapter/codex-app-server-acp-adapter.test.ts": 3423,
   "src/backend/services/session/service/chat/chat-message-handlers/handlers/load-session.handler.test.ts": 1576,
